+++ common/src/main/java/com/modeunsa/global/kafka/outbox/OutboxStore.java
@@ -11,5 +11,5 @@ public interface OutboxStore {
   void markFailed(Long id, String errorMessage, int maxRetry);
 
   // 이미 SENT 상태로 마킹된 OutboxEventView 데이터를 삭제
-  int deleteAlreadySentEventByIds(List<Long> ids);
+  long deleteAlreadySentEventByIds(List<Long> ids);
 }
